+++ assets/webpack.config.js
@@ -63,7 +63,8 @@ const populateEntries = () => {
   const styleSheets = [
     { adaptive: './styles/adaptive/adaptive-reset.scss' },
     { styles: './styles/index.scss' },
-    { preview: './styles/preview.scss' }];
+    { preview: './styles/preview.scss' },
+  ];
 
   // Merge the attributes of all found activities and the initialEntries
   // into one single object.
@@ -77,9 +78,9 @@ const populateEntries = () => {
   if (
     Object.keys(merged).length !=
     Object.keys(initialEntries).length +
-    2 * foundActivities.length +
-    2 * foundParts.length +
-    styleSheets.length
+      2 * foundActivities.length +
+      2 * foundParts.length +
+      styleSheets.length
   ) {
     throw new Error(
       'Encountered a possible naming collision in activity or part manifests. Aborting.',
@@ -319,7 +320,7 @@ module.exports = (env, options) => ({
       outputFilename: '../licenses.json',
       licenseOverrides: {
         'janus-script@1.9.2': 'MIT',
-        'phoenix_html@3.2.0': 'MIT',
+        'phoenix_html@3.3.1': 'MIT',
         'typed-function@2.0.0': 'MIT',
       },
       unacceptableLicenseTest: (licenseIdentifier) => {