+++ config/dev.exs
@@ -30,7 +30,8 @@ config :oli,
     favicons: System.get_env("BRANDING_FAVICONS_DIR", "/branding/dev/favicons")
   ],
   always_use_persistent_login_sessions:
-    get_env_as_boolean.("ALWAYS_USE_PERSISTENT_LOGIN_SESSIONS", "false")
+    get_env_as_boolean.("ALWAYS_USE_PERSISTENT_LOGIN_SESSIONS", "false"),
+  log_incomplete_requests: get_env_as_boolean.("LOG_INCOMPLETE_REQUESTS", "true")
 
 config :oli, :vendor_property,
   workspace_logo:
@@ -68,6 +69,16 @@ config :oli, :cashnet_provider,
   cashnet_client: System.get_env("CASHNET_CLIENT"),
   cashnet_gl_number: System.get_env("CASHNET_GL_NUMBER")
 
+# Configurable http/https protocol options for cowboy
+# https://ninenines.eu/docs/en/cowboy/2.5/manual/cowboy_http/
+http_max_header_name_length =
+  System.get_env("HTTP_MAX_HEADER_NAME_LENGTH", "64") |> String.to_integer()
+
+http_max_header_value_length =
+  System.get_env("HTTP_MAX_HEADER_VALUE_LENGTH", "4096") |> String.to_integer()
+
+http_max_headers = System.get_env("HTTP_MAX_HEADERS", "100") |> String.to_integer()
+
 # For development, we disable any cache and enable
 # debugging and code reloading.
 #
@@ -76,7 +87,12 @@ config :oli, :cashnet_provider,
 # with webpack to recompile .js and .css sources.
 config :oli, OliWeb.Endpoint,
   http: [
-    port: String.to_integer(System.get_env("HTTP_PORT", "80"))
+    port: String.to_integer(System.get_env("HTTP_PORT", "80")),
+    protocol_options: [
+      max_header_name_length: http_max_header_name_length,
+      max_header_value_length: http_max_header_value_length,
+      max_headers: http_max_headers
+    ]
   ],
   url: [
     scheme: System.get_env("SCHEME", "https"),
@@ -87,7 +103,12 @@ config :oli, OliWeb.Endpoint,
     port: String.to_integer(System.get_env("HTTPS_PORT", "443")),
     otp_app: :oli,
     keyfile: System.get_env("SSL_KEY_PATH", "priv/ssl/localhost.key"),
-    certfile: System.get_env("SSL_CERT_PATH", "priv/ssl/localhost.crt")
+    certfile: System.get_env("SSL_CERT_PATH", "priv/ssl/localhost.crt"),
+    protocol_options: [
+      max_header_name_length: http_max_header_name_length,
+      max_header_value_length: http_max_header_value_length,
+      max_headers: http_max_headers
+    ]
   ],
   debug_errors: true,
   code_reloader: true,