+++ config/releases.exs
@@ -103,7 +103,8 @@ config :oli,
   screen_idle_timeout_in_seconds:
     String.to_integer(System.get_env("SCREEN_IDLE_TIMEOUT_IN_SECONDS", "1800")),
   always_use_persistent_login_sessions:
-    get_env_as_boolean.("ALWAYS_USE_PERSISTENT_LOGIN_SESSIONS", "false")
+    get_env_as_boolean.("ALWAYS_USE_PERSISTENT_LOGIN_SESSIONS", "false"),
+  log_incomplete_requests: get_env_as_boolean.("LOG_INCOMPLETE_REQUESTS", "true")
 
 default_description = """
 The Open Learning Initiative enables research and experimentation with all aspects of the learning experience.
@@ -191,17 +192,26 @@ help_provider =
 
 config :oli, :help, dispatcher: help_provider
 
-# Configurable http/https max header length accepted for Cowboy. Default is 4096
+# Configurable http/https protocol options for cowboy
 # https://ninenines.eu/docs/en/cowboy/2.5/manual/cowboy_http/
+http_max_header_name_length =
+  System.get_env("HTTP_MAX_HEADER_NAME_LENGTH", "64") |> String.to_integer()
+
 http_max_header_value_length =
-  System.get_env("HTTP_MAX_VALUE_HEADER_LENGTH", "4096") |> String.to_integer()
+  System.get_env("HTTP_MAX_HEADER_VALUE_LENGTH", "4096") |> String.to_integer()
+
+http_max_headers = System.get_env("HTTP_MAX_HEADERS", "100") |> String.to_integer()
 
 config :oli, OliWeb.Endpoint,
   server: true,
   http: [
     :inet6,
     port: String.to_integer(System.get_env("HTTP_PORT", "80")),
-    protocol_options: [max_header_value_length: http_max_header_value_length]
+    protocol_options: [
+      max_header_name_length: http_max_header_name_length,
+      max_header_value_length: http_max_header_value_length,
+      max_headers: http_max_headers
+    ]
   ],
   url: [
     scheme: System.get_env("SCHEME", "https"),
@@ -218,7 +228,11 @@ if System.get_env("SSL_CERT_PATH") && System.get_env("SSL_KEY_PATH") do
       otp_app: :oli,
       keyfile: System.get_env("SSL_CERT_PATH", "priv/ssl/localhost.key"),
       certfile: System.get_env("SSL_KEY_PATH", "priv/ssl/localhost.crt"),
-      protocol_options: [max_header_value_length: http_max_header_value_length]
+      protocol_options: [
+        max_header_name_length: http_max_header_name_length,
+        max_header_value_length: http_max_header_value_length,
+        max_headers: http_max_headers
+      ]
     ]
 end